+++ packages/ui/src/components/property-editor/editor-items/BooleanItem.jsx
@@ -1,0 +1,13 @@
+import { Checkbox } from "../../checkbox";
+
+export function BooleanItem({ value, onChange, style }) {
+  return (
+    <Checkbox
+      checked={value}
+      style={style}
+      onChange={(e) => {
+        onChange && onChange(e.target.checked);
+      }}
+    />
+  );
+}