+++ packages/ui/src/components/property-editor/editor-items/ButtonItem.jsx
@@ -1,6 +1,21 @@
 import { Button } from "../../button";
 import "./ButtonItem.scss";
 
-export function ButtonItem({ item }) {
-  return <Button onClick={item.onClick}>{item.title}</Button>;
+export function ButtonItem({ label, style, item }) {
+  const { onClick, justified, iconOnly, shape, variant } = item;
+  return (
+    <Button
+      onClick={onClick}
+      style={{
+        ...style,
+        height: 26,
+      }}
+      justified={justified}
+      iconOnly={iconOnly}
+      shape={shape}
+      variant={variant}
+    >
+      {label}
+    </Button>
+  );
 }