+++ packages/ui/src/components/property-editor/editor-items/NumberInputItem.jsx
@@ -1,0 +1,18 @@
+import { InputEditor } from "../../input-editor";
+
+export function NumberInputItem({ value, item, style, onChange }) {
+  const { min = 0, max = 100 } = item;
+  return (
+    <InputEditor
+      type="number"
+      value={value}
+      min={min}
+      max={max}
+      width="100%"
+      style={style}
+      onInput={(e) => {
+        onChange && onChange(e.target.value);
+      }}
+    />
+  );
+}