+++ packages/ui/src/components/property-editor/editor-items/SelectItem.jsx
@@ -1,0 +1,12 @@
+import { Select } from "../../select";
+
+export function SelectItem({ value, item, style, onChange }) {
+  return (
+    <Select
+      value={value}
+      style={style}
+      onChange={onChange}
+      options={item.options}
+    />
+  );
+}