+++ packages/ui/src/components/property-editor/editor-items/SwitchItem.jsx
@@ -1,0 +1,15 @@
+import { Switch } from "../../switch";
+
+export function SwitchItem({ item, value, onChange, style }) {
+  const { variant } = item;
+  return (
+    <Switch
+      checked={value}
+      style={style}
+      variant={variant}
+      onChange={(e) => {
+        onChange && onChange(e.target.checked);
+      }}
+    />
+  );
+}