+++ packages/ui/src/index.js
@@ -78,3 +78,4 @@ export * from "./components/card/CardAvatar";
 export * from "./components/card/CardActions";
 export * from "./components/ghost";
 export * from "./components/blank";
+export * from "./components/select";