+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/design-tokens
 
+## 0.0.42
+
+### Patch Changes
+
+- update property editor's sub item (slider, number, select, switch, tab, boolean)
+
 ## 0.0.40
 
 ### Patch Changes